Allow an expression itself to express if it is literal
This provides a minor performance gain, as it saves a type assertion in
Eval. This is a silly way to go about performance improvements.
Fortunately this also paves the way for my next idea, which is to make
everything an expression; saving the need for the costly interface
assertion that's left in Eval.